remove @zitadel/node, move fcn to /client
