+++ apps/login/package.json
@@ -39,7 +39,6 @@
     "@tailwindcss/forms": "0.5.7",
     "@vercel/analytics": "^1.2.2",
     "@zitadel/client": "workspace:*",
-    "@zitadel/node": "workspace:*",
     "@zitadel/proto": "workspace:*",
     "clsx": "1.2.1",
     "copy-to-clipboard": "^3.3.3",
