+++ packages/zitadel-client/package.json
@@ -46,6 +46,9 @@
   "dependencies": {
     "@bufbuild/protobuf": "^2.2.2",
     "@connectrpc/connect": "^2.0.0",
+    "@connectrpc/connect-node": "^2.0.0",
+    "@connectrpc/connect-web": "^2.0.0",
+    "jose": "^5.3.0",
     "@zitadel/proto": "workspace:*"
   },
   "devDependencies": {
