+++ packages/zitadel-client/src/index.ts
@@ -5,3 +5,5 @@ export { NewAuthorizationBearerInterceptor } from "./interceptors";
 export { create, fromJson, toJson } from "@bufbuild/protobuf";
 export { TimestampSchema, timestampDate, timestampFromDate, timestampFromMs, timestampMs } from "@bufbuild/protobuf/wkt";
 export type { Duration, Timestamp } from "@bufbuild/protobuf/wkt";
+
+export { createServerTransport, newSystemToken } from "./grpc";
