--- packages/zitadel-node/.eslintrc.cjs
@@ -1,5 +1,0 @@
-module.exports = {
-  root: true,
-  // TODO: React is not used in the server package
-  extends: ["zitadel"],
-};
