--- packages/zitadel-node/package.json
@@ -1,46 +1,0 @@
-{
-  "name": "@zitadel/node",
-  "version": "0.0.0",
-  "type": "module",
-  "sideEffects": false,
-  "license": "MIT",
-  "exports": {
-    ".": {
-      "types": "./dist/index.d.ts",
-      "import": "./dist/index.js",
-      "require": "./dist/index.cjs"
-    }
-  },
-  "files": [
-    "dist/**"
-  ],
-  "publishConfig": {
-    "access": "public"
-  },
-  "scripts": {
-    "build": "tsup",
-    "test": "pnpm test:unit",
-    "test:watch": "pnpm test:unit:watch",
-    "test:unit": "vitest",
-    "test:unit:watch": "vitest --watch",
-    "dev": "tsup --watch",
-    "lint": "eslint \"src/**/*.ts*\"",
-    "clean": "rm -rf .turbo && rm -rf node_modules && rm -rf dist && rm -rf src/proto"
-  },
-  "peerDependencies": {
-    "@zitadel/client": "workspace:*",
-    "@connectrpc/connect": "^2.0.0"
-  },
-  "dependencies": {
-    "@connectrpc/connect-node": "^2.0.0",
-    "@connectrpc/connect-web": "^2.0.0",
-    "jose": "^5.3.0"
-  },
-  "devDependencies": {
-    "@connectrpc/connect": "^2.0.0",
-    "@types/node": "^22.9.0",
-    "@zitadel/client": "workspace:*",
-    "@zitadel/tsconfig": "workspace:*",
-    "eslint-config-zitadel": "workspace:*"
-  }
-}
