--- packages/zitadel-node/tsconfig.json
@@ -1,5 +1,0 @@
-{
-  "extends": "@zitadel/tsconfig/tsup.json",
-  "include": ["./src/**/*"],
-  "exclude": ["dist", "build", "node_modules"]
-}
