--- packages/zitadel-node/turbo.json
@@ -1,15 +1,0 @@
-{
-  "extends": [
-    "//"
-  ],
-  "tasks": {
-    "build": {
-      "outputs": [
-        "dist/**"
-      ],
-      "dependsOn": [
-        "@zitadel/client#build"
-      ]
-    }
-  }
-}
